fix validation test
