+++ library/src/test/java/io/voucherify/android/helper/MockSuccessCall.java
@@ -1,0 +1,45 @@
+package io.voucherify.android.helper;
+
+import java.io.IOException;
+
+import okhttp3.Request;
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
+
+public class MockSuccessCall<T> implements Call<T> {
+    @Override
+    public Response<T> execute() throws IOException {
+        return Response.success(null);
+    }
+
+    @Override
+    public void enqueue(Callback<T> callback) {
+
+    }
+
+    @Override
+    public boolean isExecuted() {
+        return false;
+    }
+
+    @Override
+    public void cancel() {
+
+    }
+
+    @Override
+    public boolean isCanceled() {
+        return false;
+    }
+
+    @Override
+    public Call<T> clone() {
+        return null;
+    }
+
+    @Override
+    public Request request() {
+        return null;
+    }
+}
